change font, change view transitions to header, change data
