+++ src/components/BaseHead.astro
@@ -1,14 +1,18 @@
 ---
 // Import the global.css file here so that it is included on
 // all pages through the use of the <BaseHead /> component.
-import '../styles/global.css';
-import { SITE_TITLE, SITE_DESCRIPTION, OG_IMAGE, SITE_KEYWORDS } from "../consts.js";
-
+import "../styles/global.css";
+import {
+  SITE_TITLE,
+  SITE_DESCRIPTION,
+  OG_IMAGE,
+  SITE_KEYWORDS,
+} from "../consts.js";
 
 interface Props {
-	title: string;
-	description: string;
-	image: string;
+  title: string;
+  description: string;
+  image: string;
   keywords: string;
 }
 
@@ -20,34 +24,36 @@ let pageImage = image;
 let pageKeywords = keywords;
 
 // Custom Titles
-let holderTitle = SITE_TITLE
+let holderTitle = SITE_TITLE;
 if (pageTitle != null) {
   holderTitle = pageTitle;
-}; 
+}
 const newTitle = holderTitle;
 
 // Custom Description
-let holderDesc = SITE_DESCRIPTION
+let holderDesc = SITE_DESCRIPTION;
 if (pageDescription != null) {
   holderDesc = pageDescription;
-};
-const newDesc = holderDesc
+}
+const newDesc = holderDesc;
 
 // Custom OG IMG
-let holderImage = OG_IMAGE
+let holderImage = OG_IMAGE;
 if (pageImage != null) {
   holderImage = pageImage;
-};
-const newImage = holderImage
+}
+const newImage = holderImage;
 
 // Custom keywords
-let holderKeywords = SITE_KEYWORDS
+let holderKeywords = SITE_KEYWORDS;
 if (pageKeywords != null) {
   holderKeywords = pageKeywords;
-};
-const newKeywords = holderKeywords
+}
+const newKeywords = holderKeywords;
 
 const canonicalURL = new URL(Astro.url.pathname, Astro.site);
+
+import { ViewTransitions } from "astro:transitions";
 ---
 
 <!-- Site Created By:    
@@ -65,9 +71,7 @@ const canonicalURL = new URL(Astro.url.pathname, Astro.site);
                                                __/ |                
                                               |___/                 
 
-    Nice to see you here.  Find new student pilots with Right Rudder Marketing.  https://rightruddermarketing.com/contact-us -->
-
-<!-- Global Metadata -->
+    Nice to see you here.  Find new student pilots with Right Rudder Marketing.  https://rightruddermarketing.com/contact-us --><!-- Global Metadata -->
 <meta charset="utf-8" />
 <meta name="viewport" content="width=device-width,initial-scale=1" />
 <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
@@ -76,6 +80,12 @@ const canonicalURL = new URL(Astro.url.pathname, Astro.site);
 <!-- Font preloads -->
 <!-- <link rel="preload" href="/fonts/atkinson-regular.woff" as="font" type="font/woff" crossorigin /> -->
 <!-- <link rel="preload" href="/fonts/atkinson-bold.woff" as="font" type="font/woff" crossorigin /> -->
+<link rel="preconnect" href="https://fonts.googleapis.com" />
+<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+<link
+  href="https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@100;200;300;400;500;600;700;800;900&display=swap"
+  rel="stylesheet"
+/>
 
 <!-- Canonical URL -->
 <link rel="canonical" href={canonicalURL} />
@@ -101,11 +111,18 @@ const canonicalURL = new URL(Astro.url.pathname, Astro.site);
 <meta property="twitter:image" content={new URL(newImage, Astro.url)} />
 
 <!-- Google tag (gtag.js) -->
-<script type="text/partytown" async src="https://www.googletagmanager.com/gtag/js?id=G-84N22Y2BJ1"></script>
+<script
+  type="text/partytown"
+  async
+  src="https://www.googletagmanager.com/gtag/js?id=G-84N22Y2BJ1"></script>
 <script type="text/partytown">
-	window.dataLayer = window.dataLayer || [];
-	function gtag(){dataLayer.push(arguments);}
-	gtag('js', new Date());
+  window.dataLayer = window.dataLayer || [];
+  function gtag() {
+    dataLayer.push(arguments);
+  }
+  gtag("js", new Date());
 
-	gtag('config', 'G-84N22Y2BJ1');
+  gtag("config", "G-84N22Y2BJ1");
 </script>
+
+<ViewTransitions />
