+++ src/components/ContactUs.astro
@@ -12,7 +12,7 @@ import ContactUsForm from "../components/ContactUsForm.astro";
 
 <section
   class="relative bg-two-cessna bg-center bg-cover bg-fixed bg-main-red text-white"
-  id="contact-us"
+  id="contactUs"
 >
   <!-- bg-overlay to darken -->
   <div class="absolute bg-main-red/60 w-full h-full z-0"></div>
