+++ src/components/ContactUsForm.astro
@@ -2,24 +2,52 @@
 
 ---
 
-<form class="mx-auto max-w-xl pb-10" id="contact_form"  method="POST" onsubmit="validateForm(this);">
+<form
+  class="mx-auto max-w-xl pb-10"
+  method="POST"
+  onsubmit="validateForm(this);"
+>
   <div class="grid grid-cols-1 gap-x-8 gap-y-2 sm:grid-cols-2">
     <div>
       <!-- <label for="first-name" class="block text-sm leading-6 text-white">First name</label> -->
       <div class="mt-2.5">
-        <input type="text" name="first-name" id="first-name" autocomplete="given-name" placeholder="First Name" required class="block w-full py-2 text-white shadow-sm ring-inset ring-gray-300 placeholder:text-gray-300 bg-transparent border-b-[1px] border-gray-300 sm:text-sm sm:leading-6 outline-none">
+        <input
+          type="text"
+          name="first-name"
+          id="first-name"
+          autocomplete="given-name"
+          placeholder="First Name"
+          required
+          class="block w-full py-2 text-white shadow-sm ring-inset ring-gray-300 placeholder:text-gray-300 bg-transparent border-b-[1px] border-gray-300 sm:text-sm sm:leading-6 outline-none"
+        />
       </div>
     </div>
     <div>
       <!-- <label for="last-name" class="block text-sm leading-6 text-white">Last name</label> -->
       <div class="mt-2.5">
-        <input type="text" name="last-name" id="last-name" autocomplete="family-name" placeholder="Last Name" required class="block w-full py-2 text-white shadow-sm ring-inset ring-gray-300 placeholder:text-gray-300 bg-transparent border-b-[1px] border-gray-300 sm:text-sm sm:leading-6 outline-none">
+        <input
+          type="text"
+          name="last-name"
+          id="last-name"
+          autocomplete="family-name"
+          placeholder="Last Name"
+          required
+          class="block w-full py-2 text-white shadow-sm ring-inset ring-gray-300 placeholder:text-gray-300 bg-transparent border-b-[1px] border-gray-300 sm:text-sm sm:leading-6 outline-none"
+        />
       </div>
     </div>
     <div class="sm:col-span-2">
       <!-- <label for="email" class="block text-sm leading-6 text-white">Email</label> -->
       <div class="mt-2.5">
-        <input type="email" name="email" id="email" autocomplete="email" placeholder="Email Address" required class="block w-full py-2 text-white shadow-sm ring-inset ring-gray-300 placeholder:text-gray-300 bg-transparent border-b-[1px] border-gray-300 sm:text-sm sm:leading-6 outline-none">
+        <input
+          type="email"
+          name="email"
+          id="email"
+          autocomplete="email"
+          placeholder="Email Address"
+          required
+          class="block w-full py-2 text-white shadow-sm ring-inset ring-gray-300 placeholder:text-gray-300 bg-transparent border-b-[1px] border-gray-300 sm:text-sm sm:leading-6 outline-none"
+        />
       </div>
     </div>
     <p class="ohno">
@@ -30,14 +58,29 @@
     <div class="sm:col-span-2">
       <!-- <label for="phone-number" class="block text-sm leading-6 text-white">Phone number</label> -->
       <div class="relative mt-2.5">
-        <input type="tel" name="phone-number" id="phone-number" autocomplete="tel" placeholder="Phone Number" required class="block w-full py-2 text-white shadow-sm ring-inset ring-gray-300 placeholder:text-gray-300 bg-transparent border-b-[1px] border-gray-300 sm:text-sm sm:leading-6 outline-none">
+        <input
+          type="tel"
+          name="phone-number"
+          id="phone-number"
+          autocomplete="tel"
+          placeholder="Phone Number"
+          required
+          class="block w-full py-2 text-white shadow-sm ring-inset ring-gray-300 placeholder:text-gray-300 bg-transparent border-b-[1px] border-gray-300 sm:text-sm sm:leading-6 outline-none"
+        />
       </div>
     </div>
     <div class="sm:col-span-2">
       <!-- <label for="message" class="block text-sm leading-6 text-white">How can we help?</label> -->
       <div class="mt-2.5">
         <!-- Need to get align bottom... -->
-        <textarea name="message" id="message" rows="4" placeholder="How we help?" required class="block w-full py-2 text-white shadow-sm ring-inset ring-gray-300 placeholder:text-gray-300 bg-transparent border-b-[1px] border-gray-300 sm:text-sm sm:leading-6 outline-none"></textarea>
+        <textarea
+          name="message"
+          id="message"
+          rows="4"
+          placeholder="How we help?"
+          required
+          class="block w-full py-2 text-white shadow-sm ring-inset ring-gray-300 placeholder:text-gray-300 bg-transparent border-b-[1px] border-gray-300 sm:text-sm sm:leading-6 outline-none"
+        ></textarea>
       </div>
     </div>
     <div class="flex gap-x-4 sm:col-span-2">
@@ -52,18 +95,26 @@
           aria-checked="false"
           aria-labelledby="switch-1-label"
         >
-          <span class="sr-only">I agree to terms & conditions provided by the company. By providing my phone number, I agree to receive text messages from the business.</span>
+          <span class="sr-only"
+            >I agree to terms & conditions provided by the company. By providing
+            my phone number, I agree to receive text messages from the business.</span
+          >
           <!-- Enabled: "translate-x-3.5", Not Enabled: "translate-x-0" -->
           <span
-              x-bind:class="{ 'translate-x-3.5': isChecked, 'translate-x-0': !isChecked }"
-              aria-hidden="true"
-              class="h-4 w-4 transform rounded-full bg-white shadow-sm ring-1 ring-gray-900/5 transition duration-200 ease-in-out"
+            x-bind:class="{ 'translate-x-3.5': isChecked, 'translate-x-0': !isChecked }"
+            aria-hidden="true"
+            class="h-4 w-4 transform rounded-full bg-white shadow-sm ring-1 ring-gray-900/5 transition duration-200 ease-in-out"
           ></span>
         </button>
-        <input type="hidden" x-bind:value="isChecked ? 'yes' : 'no'" name="agree-terms">
+        <input
+          type="hidden"
+          x-bind:value="isChecked ? 'yes' : 'no'"
+          name="agree-terms"
+        />
       </div>
       <label class="text-sm leading-6 text-gray-200" id="switch-1-label">
-        I agree to <a href="">terms & conditions</a> provided by the company. By providing my phone number, I agree to receive text messages from the business.
+        I agree to <a href="">terms & conditions</a> provided by the company. By
+        providing my phone number, I agree to receive text messages from the business.
       </label>
     </div>
   </div>
@@ -73,19 +124,18 @@
 </form>
 
 <script>
-
-(window as any).validateForm = validateForm;
-function validateForm(form) {
-  var x = document.forms["contact_form"]["confirm-email"].value;
-  if ( x == "" || x == null ){ // if the honeypot was ignored, it's a hu-mon
-    form.action="https://usebasin.com/f/d5eca44afead"; // link to process form and redirect to thank you
-  }
-  else{ // the honeypot was filled in, it's a robot
-    form.action="https://flyredarrow.com/contact-confirmation" // link directly to thank you without actually processing form    
-    return false;
+  (window as any).validateForm = validateForm;
+  function validateForm(form) {
+    var x = document.forms["contact_form"]["confirm-email"].value;
+    if (x == "" || x == null) {
+      // if the honeypot was ignored, it's a hu-mon
+      form.action = "https://usebasin.com/f/d5eca44afead"; // link to process form and redirect to thank you
+    } else {
+      // the honeypot was filled in, it's a robot
+      form.action = "https://flyredarrow.com/contact-confirmation"; // link directly to thank you without actually processing form
+      return false;
     }
-}
-
+  }
 </script>
 
 <style>
