+++ src/components/DiscoveryFlightCTA.astro
@@ -11,7 +11,7 @@
       <h2 class="text-3xl font-semibold text-center tracking-tight">
         Book a Discovery Flight
       </h2>
-      <h3 class="pt-10 text-xl italic">
+      <h3 class="pt-10 text-xl italic text-center">
         Schedule with us today to learn more about the pilot training process
         and see if the world of flight is for you.
       </h3>
