+++ src/components/TeamMemberCard.astro
@@ -9,7 +9,7 @@ const { teamMember } = Astro.props;
     class="flex flex-col rounded-lg overflow-hidden gap-3 max-w-md lg:w-64 lg:min-w-64"
   >
     <img
-      src={teamMember.image.src}
+      src={teamMember.image}
       class="aspect-square object-cover w-full rounded-lg border-4 border-gray-700 bg-white"
       alt={teamMember.alt}
     />
