+++ src/data/team.js
@@ -1,0 +1,56 @@
+export const teamMembers = [
+  {
+    name: "Tomas Peralta",
+    title: "Chief Flight Instructor | Founder",
+    subtitle:
+      "Certificates & Ratings: ATP, FAA Designated Pilot Examiner, Commercial ASEL, AMEL, CFI, CFII",
+    bio: [
+      "Tomas Peralta is the Chief Flight Instructor at Red Arrow Flight Academy, bringing with him over 20 years of experience and an extensive knowledge of aviation. Mr. Peralta fostered a love for flying, thanks in part to his family’s involvement in aviation. His dad, mom, and sister all hold private pilot certificates, so he began his training in 2001 while attending college. He went on to earn his Bachelor of Science in Industrial Engineering from ITESM in Monterrey, Mexico, as well as a Master of Science in Industrial Engineering and an MBA from Georgia Institute of Technology.",
+      "After returning to the El Paso area, Mr. Peralta became a founding member of Red Arrow Flight Academy. He holds an ATP single-engine land certificate, a commercial multi-engine aircraft certificate, and a Cirrus Vision Jet Type Rating. In addition, he is a CFI and CFII. Mr. Peralta has flown and owned a variety of aircraft for personal, training, and corporate flying purposes. Mr. Peralta’s extensive experience and expertise have earned him roles as both a FAASTeam Representative and an ABQ FSDO Designated Pilot Examiner.",
+      "He is also a dedicated warbird enthusiast and serves as a board member at the War Eagles Air Museum, where he shares his passion for aviation history. As an active volunteer, he has assisted with the maintenance and operation of a TF-51 Mustang. Mr. Peralta is also the President of the EAA (Experimental Aircraft Association) 1570 Chapter. In the wintertime, Mr. Peralta is a certified National Ski Patroller and Outdoor Emergency Care instructor. He is often accompanied by his trusty German Shepard sidekick, Belka.",
+    ],
+    image: "/img/our-team/tomas-peralta.webp",
+    alt: "Founder and Owner Tomas Peralta at Red Arrow Flight Academy",
+    email: "tomas@flyredarrow.com",
+    phone: "(575) 201-6160 ext. 702",
+  },
+  {
+    name: "Elsa Peralta",
+    title: "Administrative Manager | Founder",
+    subtitle: "Certificates & Ratings: Private Pilot Certificate, ASEL",
+    bio: [
+      "Elsa carries the torch of her family's passion for flying. Drawing from her rich experiences, she has become a driving force behind the success of Red Arrow Flight Academy. Currently, Elsa holds a Private Pilot Certificate (PPC) for Airplane Single-Engine Land (ASEL), and her thirst for knowledge and skill in aviation continues to soar. She's set her sights on obtaining her instrument rating, an ambition that will further enrich her aviation journey.",
+      "Elsa's roots extend into the automotive industry. Her career in quality control, project management, logistics, and general administration has honed her skills in managing complex operations and teams. Her journey of academic excellence culminated in a B.S. in Industrial Engineering from ITESM, a master's in international business from Pepperdine University, and an MBA from EGADE.",
+      "One of Elsa's assignments involved a pivotal role at Daimler AG, the parent company of the prestigious Mercedes-Benz. Here, she oversaw the integration of the supply chain system during Daimler AG's merger with Chrysler, showcasing her ability to handle high-stakes endeavors on a global scale. Elsa's multifaceted journey, blending aviation passion with astute leadership and a diverse skill set, exemplifies the power of determination and dedication. Her unwavering commitment to excellence continues to chart new horizons for herself and the institutions she serves.",
+    ],
+    image: "/img/our-team/elsa-peralta.webp",
+    alt: "Founder and Owner Elsa Peralta at Red Arrow Flight Academy",
+    email: "elsa@flyredarrow.com",
+    phone: "(575) 201-6160 ext. 705",
+  },
+  {
+    name: "Jazmine Anaya",
+    title: "Dispatch Manager",
+    bio: [
+      "El Paso native Jazmine Anaya is a born adventurer with an insatiable passion for exploration. Born and raised in El Paso, Texas, Jazmine's heart has always been captivated by the allure of distant horizons and new frontiers.  In her pursuit of wanderlust, Jazmine has taken a remarkable leap into the world of aviation by becoming an integral part of the Red Arrow team as a dispatcher. Here, her dedication and enthusiasm for the skies have found a perfect platform to flourish.",
+      "Jazmine's journey in aviation is marked by her tenacity and ambition. She is a proud recipient of the EAA Chapter 1570 flight training scholarship, a recognition of her unwavering commitment to taking flight. As a student pilot, she is in the process of transforming her dreams of flight into a soaring reality.  Her introduction to the world of aviation represents a pivotal moment in her life, one that promises to open up a world of opportunities and adventures. Jazmine's journey is a testament to the power of following one's passion, even when the path may seem distant and uncertain.",
+      "As she continues to explore the boundless skies, Jazmine Anaya serves as an inspiring example of someone who dares to chase their dreams and embrace the thrilling unknown of aviation. Her story is one of determination, ambition, and the endless possibilities that the world of flight has to offer.",
+    ],
+    image: "/img/our-team/jazmine-anaya.webp",
+    alt: "Dispatch Manager Jazmine Anaya at Red Arrow Flight Academy",
+    email: "janaya@flyredarrow.com",
+    phone: "(575) 201-6160 ext. 701",
+  },
+  {
+    name: "Danielle Nelson",
+    title: "Dispatch",
+    bio: [
+      "Danielle's journey into the world of aviation is a testament to the power of dreams and determination. Graduating high school just this year, she wasted no time in taking the plunge into the aviation industry, making a resounding impact at Red Arrow.  Her journey towards the skies began when she met a U.S. Army helicopter pilot during her junior year. This fateful encounter ignited a spark within her, redirecting her life's course. Initially, Danielle had aspirations to pursue a career in the medical field, but the allure of aviation proved irresistible.",
+      "As a student pilot at Red Arrow, Danielle is taking her first steps toward realizing her dream of becoming a pilot. Her determination to learn and excel in the world of aviation is nothing short of inspiring. She is driven by a deep-seated passion for flight, and there's no looking back on the path she's chosen. Danielle has set her sights on an ambitious goal - to become an instructor right here at Red Arrow once she completes her flight training. Her commitment to passing on her knowledge and love for aviation to future pilots is a testament to her dedication.",
+      "While the exact direction her aviation journey will take her after instruction remains undecided, one thing is certain: Danielle wants to continue flying. Her desire and motivation to achieve her goals shine brightly, and Red Arrow is proud to have her as an integral part of the team. Danielle represents the embodiment of dreams taking flight, and her journey is a source of inspiration for all who share her passion for aviation.",
+    ],
+    image: "/img/our-team/danielle-nelson.webp",
+    alt: "Dispatch Manager Danielle Nelson at Red Arrow Flight Academy",
+    email: "dnelson@flyredarrow.com",
+  },
+];
