+++ src/pages/about/our-team.astro
@@ -3,15 +3,15 @@ import DiscoveryFlightCTA from "../../components/DiscoveryFlightCTA.astro";
 import ContactUs from "../../components/ContactUs.astro";
 import BaseLayout from "../../layouts/BaseLayout.astro";
 import TeamMemberCard from "../../components/TeamMemberCard.astro";
-import { teamMembers } from "../../consts";
+import { teamMembers } from "../../data/team.js";
 import Location from "../../components/Location.astro";
 ---
 
-<BaseLayout 
-  pageTitle = "Our Team | Red Arrow Flight Academy" 
-  pageDescription = "Meet our expert and caring team ready to take you to new heights. Get to know our leadership and dispatch teams at Red Arrow. See who you’ll be flying with!"
-  pageKeywords = "Team, staff, Red Arrow, leadership, dispatch, CFII, CFI, pilot, training"
-  >
+<BaseLayout
+  pageTitle="Our Team | Red Arrow Flight Academy"
+  pageDescription="Meet our expert and caring team ready to take you to new heights. Get to know our leadership and dispatch teams at Red Arrow. See who you’ll be flying with!"
+  pageKeywords="Team, staff, Red Arrow, leadership, dispatch, CFII, CFI, pilot, training"
+>
   <header
     class="relative h-screen-3/4 bg-our-team bg-cover bg-center bg-black flex flex-col items-center justify-center"
   >
